pyproject.toml
[tool.poetry]
name = "powertools-oas-validator"
version = "0.1.0"
description = ""
authors = ["Rasmus Hansen <Rasmus.Hansen@LifeWorks.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "powertools_oas_validator"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-lambda-powertools = "^2.18.0"
jmespath = "^1.0.1"
openapi-spec-validator = "^0.5.7"
chocs-middleware-openapi = "^1.2.2"
fastjsonschema = "^2.17.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.4.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
autoflake = "^2.2.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "--cov=powertools_oas_validator --cov-report term-missing"
